requirements.txt
Flask
Flask-SQLAlchemy
requests
beautifulsoup4
lxml
Werkzeug
MarkupSafe